+++ server/index.js
@@ -40,8 +40,8 @@ app.use(session(
 
 // use routes
 app.use('/app/login', require('./routes/login.js'));
-app.use('/app/token-exchange', require('./routes/token-exchange.js'));
-app.use('/app/token-refresh', require('./routes/token-refresh.js'));
+app.use('/app/callback', require('./routes/callback.js'));
+app.use('/app/refresh', require('./routes/refresh.js'));
 app.use('/app/logout', require('./routes/logout.js'));
 app.use('/app/register', require('./routes/register.js'));
 app.use('/app/me', require('./routes/me.js'));