+++ server/routes/refresh.js
@@ -6,7 +6,7 @@ const cookie = require('../cookie.js');
 const router = express.Router();
 
 router.post('/', (req, res) => {
-    if (req.cookies['refresh_token']) {
+    if (req.cookies['app.rt']) {
         request(
             {
               method: 'POST',
@@ -15,8 +15,8 @@ router.post('/', (req, res) => {
                 'client_id': config.clientID,
                 'client_secret': config.clientSecret,
                 'grant_type': 'refresh_token',
-                'refresh_token': req.cookies['refresh_token'],
-                'access_token': req.cookies['access_token'],
+                'refresh_token': req.cookies['app.rt'],
+                'access_token': req.cookies['app.at'],
               },
               headers: {
                 'Content-Type': 'application/x-www-form-urlencoded'
@@ -27,12 +27,12 @@ router.post('/', (req, res) => {
               const { access_token, id_token, refresh_token, expires_in } = JSON.parse(body);
               if (access_token && refresh_token) {
 
-                cookie.setSecure(res, 'access_token', access_token);
-                cookie.setSecure(res, 'refresh_token', refresh_token);
+                cookie.setSecure(res, 'app.at', access_token);
+                cookie.setSecure(res, 'app.rt', refresh_token);
     
                 const expires_in_ms = expires_in * 1000;
-                cookie.setReadable(res, 'access_token_expires', Date.now() + expires_in_ms, expires_in_ms);
-                cookie.setReadable(res, "id_token", id_token);
+                cookie.setReadable(res, 'app.at_exp', Date.now() + expires_in_ms, expires_in_ms);
+                cookie.setReadable(res, "app.idt", id_token);
     
                 res.sendStatus(204);
               } else {